Extract updateGuide and updateGuideMessage from service.js to guide.js
