+++ src/discordBot/services/service.js
@@ -2,8 +2,7 @@ const axios = require("axios");
 const fs = require("fs");
 const path = require("path");
 const { logError } = require("./logger");
-const { findCoursesFromDb, findAllCourseNames, findCourseFromDb } = require("../../db/services/courseService");
-const { findCourseMemberCount } = require("../../db/services/courseMemberService");
+const { findAllCourseNames, findCourseFromDb } = require("../../db/services/courseService");
 const { courseAdminRole, facultyRole } = require("../../../config.json");
 
 require("dotenv").config();
@@ -270,92 +269,6 @@ const updateInviteLinks = async (guild) => {
   }));
 };
 
-const updateGuide = async (guild, models) => {
-  const channel = guild.channels.cache.find(
-    (c) => c.name === GUIDE_CHANNEL_NAME,
-  );
-
-  if (!channel) {
-    console.error("Guide channel not found!");
-    return;
-  }
-
-  const messages = await channel.messages.fetch({ limit: 100 });
-  const sortedMessages = messages.sort((a, b) => a.createdTimestamp - b.createdTimestamp);
-  const infoMessage = sortedMessages.first();
-
-  if (!infoMessage) {
-    console.error("No info message found! Delete the guide channel and launch the bot again to init the message.");
-    return;
-  }
-
-  await updateGuideMessage(infoMessage, sortedMessages, channel, models);
-};
-
-const updateGuideMessage = async (infoMessage, sortedMessages, channel, models) => {
-  const courseData = await findCoursesFromDb("code", models.Course, false);
-
-  const courseMemberCounts = await Promise.all(courseData.map(course =>
-    findCourseMemberCount(course.id, models.CourseMember)
-  ));
-
-  const rows = courseData.map((course, index) => {
-    const regExp = /[^0-9]*/;
-    const fullname = course.fullName;
-    const name = course.name;
-    const matches = regExp.exec(course.code)?.[0];
-    const code = matches ? matches + course.code.slice(matches.length) : course.code;
-    const count = courseMemberCounts[index];
-    return `${name} - ${code} - ${fullname} 👥${count}`;
-  });
-
-  const infoContent = `
-Käytössäsi on seuraavia komentoja:
-  - \`/join\` jolla voit liittyä kurssille
-  - \`/leave\` jolla voit poistua kurssilta
-Kirjoittamalla \`/join\` tai \`/leave\` botti antaa listan kursseista.
-
-You have the following commands available:
-  - \`/join\` which you can use to join a course
-  - \`/leave\` which you can use to leave a course
-The bot gives a list of the courses if you type \`/join\` or \`/leave\`.
-
-In course specific channels you can also list instructors with the command \`/instructors\`
-
-See more with \`/help\` command.
-
-Invitation link for the server ${invite_url}
-`;
-
-  const messagesArray = Array.from(sortedMessages.values());
-
-  const courseMessages = messagesArray.filter(m => m.id !== infoMessage.id && m.type !== 'CHANNEL_PINNED_MESSAGE');
-
-
-  await infoMessage.edit(infoContent);
-
-  const editOrSendPromises = [];
-  for (let i = 0; i < rows.length; i++) {
-    const rowContent = rows[i];
-    const courseMessage = courseMessages[i];
-
-    if (courseMessage) {
-      editOrSendPromises.push(courseMessage.edit(rowContent));
-    } else {
-      editOrSendPromises.push(
-        channel.send(rowContent).then(msg => msg.react("👤"))
-      );
-    }
-  }
-
-  await Promise.all(editOrSendPromises);
-
-  if (courseMessages.length > rows.length) {
-    const deletePromises = courseMessages.slice(rows.length).map(msg => msg.delete());
-    await Promise.all(deletePromises);
-  }
-};
-
 const isCourseCategory = async (channel, Course) => {
   if (channel && channel.name) {
     const course = await findCourseFromDb(getCourseNameFromCategory(channel.name), Course);
@@ -514,8 +427,6 @@ module.exports = {
   setEmojisUnlock,
   setEmojisHide,
   setEmojisUnhide,
-  updateGuide,
-  updateGuideMessage,
   setCoursePositionABC,
   isCourseCategory,
 };
